use crypto and rename releaseCacheManager to cacheManager
